feat(node): expand lockstep harness, supply chain certification, staking governance, and cross-subsystem refinements
lockstep_harness.rs (+69 lines): expand deterministic lockstep
execution harness with improved step tracking and divergence
detection

supply_chain/certification.rs (+70 lines): add certificate chain
validation and revocation checking for supply chain attestations

staking_governance.rs (+28 lines): expand governance stake
validation and voting power computation

perf_budget_guard.rs (+17 lines): add budget tracking refinements

isolation_rail_router.rs (+13 lines): expand rail routing logic

audience_token.rs: continued test variable and ID cleanup from
prior session (consistent naming with descriptive IDs)

cancellation_protocol.rs: minor adjustments
dgis_migration_gate.rs: migration gate refinements
compat_gates.rs: compatibility gate adjustments
divergence_gate.rs, evidence_replay_gate.rs, fork_detection.rs,
key_role_separation.rs: consistent assert! pattern in test match arms
adversarial_runner.rs, impossible_default.rs: security module tweaks

+++ crates/franken-node/src/migration/dgis_migration_gate.rs
@@ -29,18 +29,18 @@ pub struct GraphHealthSnapshot {
 pub struct HealthDelta {
     /// Positive values indicate increased risk relative to baseline.
     pub cascade_risk_delta: f64,
-    pub new_fragility_findings: i32,
-    pub new_articulation_points: i32,
+    pub new_fragility_findings: i64,
+    pub new_articulation_points: i64,
 }
 
 impl HealthDelta {
     pub fn between(baseline: GraphHealthSnapshot, projected: GraphHealthSnapshot) -> Self {
         Self {
             cascade_risk_delta: projected.cascade_risk - baseline.cascade_risk,
-            new_fragility_findings: projected.fragility_findings as i32
-                - baseline.fragility_findings as i32,
-            new_articulation_points: projected.articulation_points as i32
-                - baseline.articulation_points as i32,
+            new_fragility_findings: i64::from(projected.fragility_findings)
+                - i64::from(baseline.fragility_findings),
+            new_articulation_points: i64::from(projected.articulation_points)
+                - i64::from(baseline.articulation_points),
         }
     }
 }
@@ -133,7 +133,7 @@ fn evaluate_policy(
         });
     }
 
-    if delta.new_fragility_findings > thresholds.max_new_fragility_findings as i32 {
+    if delta.new_fragility_findings > i64::from(thresholds.max_new_fragility_findings) {
         reasons.push(RejectionReason {
             code: "DGIS-MIGRATE-FRAGILITY-DELTA".to_string(),
             detail: format!(
@@ -143,7 +143,7 @@ fn evaluate_policy(
         });
     }
 
-    if delta.new_articulation_points > thresholds.max_new_articulation_points as i32 {
+    if delta.new_articulation_points > i64::from(thresholds.max_new_articulation_points) {
         reasons.push(RejectionReason {
             code: "DGIS-MIGRATE-ARTICULATION-DELTA".to_string(),
             detail: format!(
